ai/requirements.txt
# VIBRANIUM AI Requirements
# Python 3.8+ required

# FastAPI Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0

# Lightweight analytics (no local model loading)
numpy>=2.0.0

# HTTP & API
httpx>=0.28.0
python-multipart>=0.0.12
requests>=2.32.3
urllib3>=2.0.0
idna>=3.7

# Data collection
praw>=7.7.1
google-play-scraper>=1.2.7

# LLM provider
groq>=0.11.0

# Google Sheets
gspread>=6.1.2
google-auth>=2.34.0

# Scheduling
apscheduler>=3.10.4

# Environment & Configuration
python-dotenv>=1.0.1

# Logging & Monitoring
python-json-logger>=2.0.7

# Testing (optional)
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Code Quality (optional)
black>=24.8.0
flake8>=7.1.0
